refact: split code in functions to use on tests
